+++ DeanzaClassRegister/DeanzaClassRegister/Controller/detailViewController.swift
