+++ DeanzaClassRegister/DeanzaClassRegister/Model/Courses.swift
@@ -1,0 +1,14 @@
+//
+//  Courses.swift
+//  DeanzaClassRegister
+//
+//  Created by Alvin Lin on 2018/8/26.
+//  Copyright © 2018 Alvin Lin. All rights reserved.
+//
+
+import Foundation
+
+struct Courses: Decodable {
+    let total: Int?
+    var data: [Data]
+}
