+++ DeanzaClassRegister/DeanzaClassRegister/Model/Courses2D.swift
@@ -1,0 +1,16 @@
+//
+//  Course.swift
+//  DeanzaClassRegister
+//
+//  Created by Alvin Lin on 2018/8/20.
+//  Copyright © 2018 Alvin Lin. All rights reserved.
+//
+
+import Foundation
+
+struct Courses2D {
+    var total: Int?
+    var data: [[Data]]
+    var departmentList: [String]
+    var isExpanded: [Bool]
+}
