+++ DeanzaClassRegister/DeanzaClassRegister/Model/SectionInfo.swift
